Determined if data was tidy and made changes to ensure it was
I noticed that the sessions.csv data contained 4 columns of time data. The start_time and end_time, and original_start_time and original_end_time, so I selected for the readable times which were the start_time and end_time columns. The players.csv data was tidy so no changes needed.